research/propose_agent/requirements.txt
openai>=1.0.0
langgraph>=0.0.20
python-dotenv>=1.0.0
